Divide ExtendedAppInterface
